+++ integrations/aspnetcore/src/Scalar.AspNetCore/Converters/PkceJsonConverter.cs
@@ -5,11 +5,9 @@ namespace Scalar.AspNetCore;
 
 internal sealed class PkceJsonConverter : JsonConverter<Pkce>
 {
-    public override Pkce Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
+    public override Pkce Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) =>
         // We don't have to implement this method because we don't need to deserialize the Pkce enum.
-        return default;
-    }
+        default;
 
     public override void Write(Utf8JsonWriter writer, Pkce value, JsonSerializerOptions options)
     {