+++ integrations/aspnetcore/tests/Scalar.AspNetCore.Tests/ScalarOptionsExtensionsTests.cs
@@ -58,8 +58,10 @@ public class ScalarOptionsExtensionsTests
             .AddHeaderContent("<h2>bar</h2>")
             .AddDocument("v1", "Version 1")
             .AddDocuments("v2", "v3")
+            .AddDocuments(new ScalarDocument("v4"))
             .WithBaseServerUrl("https://example.com")
-            .WithDynamicBaseServerUrl();
+            .WithDynamicBaseServerUrl()
+            .WithJavaScriptConfiguration("/scalar/config.js");
 
         // Assert
         options.Title.Should().Be("My title");
@@ -102,9 +104,10 @@ public class ScalarOptionsExtensionsTests
         options.HideClientButton.Should().BeTrue();
         options.HeadContent.Should().Be("<meta name=\"foo\" content=\"bar\"/><meta name=\"bar\" content=\"foo\"/>");
         options.HeaderContent.Should().Be("<h1>foo</h1><h2>bar</h2>");
-        options.Documents.Should().HaveCount(3).And.Contain(x => x.Title == "Version 1");
+        options.Documents.Should().HaveCount(4).And.Contain(x => x.Title == "Version 1");
         options.BaseServerUrl.Should().Be("https://example.com");
         options.DynamicBaseServerUrl.Should().BeTrue();
+        options.JavaScriptConfiguration.Should().Be("/scalar/config.js");
 
 #pragma warning restore CS0618 // Type or member is obsolete
     }