+++ packages/api-client/src/libs/local-storage.ts
@@ -81,3 +81,12 @@ export const loadAllResources = (mutators: WorkspaceStore) => {
     console.error(e)
   }
 }
+
+/**
+ * localStorage keys for all client resources
+ * to ensure we do not have any conflicts
+ */
+export const CLIENT_LS_KEYS = {
+  AUTH: 'scalar-client-auth',
+  SELECTED_SECURITY_SCHEMES: 'scalar-client-selected-security-schemes',
+} as const