+++ packages/api-reference/src/components/Section/SectionHeader.vue
@@ -32,7 +32,7 @@ defineProps<{
 }
 
 .section-header {
-  font-size: var(--font-size, var(--scalar-heading-2));
+  font-size: var(--font-size, var(--scalar-heading-1));
   font-weight: var(--font-weight, var(--scalar-bold));
   /* prettier-ignore */
   color: var(--scalar-color-1);